requirements.txt
certifi==2026.4.22
charset-normalizer==3.4.7
contourpy==1.3.3
cycler==0.12.1
filelock==3.29.0
fonttools==4.62.1
fsspec==2026.4.0
idna==3.14
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
packaging==26.2
pillow==10.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
requests==2.34.0
six==1.17.0
sympy==1.14.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.7.0
insightface==0.7.3
onnxruntime==1.17.3